WIP: Migrate to local Whisper service with multi-language caption support
This commit adds infrastructure and improvements for local Whisper transcription,
moving away from OpenAI API dependency. Includes Docker service setup, HTTP client
implementation, and enhanced multi-language caption handling.

## Completed Features

### Docker Service Integration
- Add Whisper service to docker-compose.yml using onerahmet/whisper-asr-webservice
- Configure with base model, health checks, and model caching
- Service runs on port 9000 with 4G memory allocation

### HTTP Whisper Client
- Implement new HTTP-based Whisper client (http_client.go)
- Supports audio transcription via REST API
- Handles VTT and SRT caption formatting
- Compatible with existing Client interface

### Configuration Updates
- Add WHISPER_API_URL config parameter
- Update whisper client factory to support HTTP provider
- Add validation for HTTP API URL
- Maintain backward compatibility with whisper.cpp

### Multi-Language Caption Support (CRITICAL IMPROVEMENT)
- Fix caption regeneration to preserve multiple languages
- When regenerating English, Spanish captions remain untouched
- When regenerating Spanish, English captions remain untouched
- Auto-detect mode only replaces detected language after transcription
- Prevents accidental deletion of unrelated language captions

### Comprehensive Testing
- Add service_test.go with extensive test coverage
- Test language-specific caption deletion
- Test multi-language preservation during regeneration
- Test auto-detect language handling
- Test job creation and status workflows

### Bug Fixes
- Fix int to float64 type conversion in local_client.go timestamps

## Known Issues (Requires Follow-up PR)

### Type Compatibility Issues
The caption generation service has type mismatches with domain model:
- Video.ID is string, not uuid.UUID
- Video.Status instead of Video.ProcessingStatus
- Video.Language is string, not *string
- storage.OriginalVideoPath doesn't exist as package function

These issues cause compilation errors that need to be fixed in service.go.
See LOCAL_WHISPER_MIGRATION_PROGRESS.md for detailed fix instructions.

## Environment Variables

```bash
ENABLE_CAPTION_GENERATION=true
WHISPER_PROVIDER=local
WHISPER_API_URL=http://whisper:9000
WHISPER_MODEL_SIZE=base
CAPTION_GENERATION_WORKERS=2
```

## Testing Multi-Language Scenarios

1. Generate English caption -> succeeds
2. Generate Spanish caption (English exists) -> Spanish added, English preserved
3. Regenerate English (both exist) -> English replaced, Spanish preserved
4. Auto-detect language -> only detected language replaced

## Next Steps

1. Fix type compatibility issues in captiongen/service.go
2. Update service to use storage.Paths struct methods
3. Run full test suite and fix any failures
4. Test end-to-end with Docker Whisper service
5. Update CI/CD pipeline for Whisper dependency

Files Changed:
- docker-compose.yml (add whisper service)
- internal/config/config.go (add WHISPER_API_URL)
- internal/whisper/client.go (support HTTP provider)
- internal/whisper/http_client.go (NEW - HTTP client implementation)
- internal/whisper/local_client.go (fix type conversion)
- internal/usecase/captiongen/service.go (multi-language logic)
- internal/usecase/captiongen/service_test.go (NEW - comprehensive tests)
- LOCAL_WHISPER_MIGRATION_PROGRESS.md (NEW - progress documentation)

Related Issue: Local Whisper integration for caption generation

+++ internal/whisper/local_client.go
@@ -83,8 +83,8 @@ func (c *localClient) Transcribe(ctx context.Context, audioPath string, targetLa
 	for i, segment := range whisperOutput.Transcription {
 		result.Segments = append(result.Segments, TranscriptionSegment{
 			Index:      i,
-			Start:      segment.Timestamps.From / 100.0, // Convert centiseconds to seconds
-			End:        segment.Timestamps.To / 100.0,
+			Start:      float64(segment.Timestamps.From) / 100.0, // Convert centiseconds to seconds
+			End:        float64(segment.Timestamps.To) / 100.0,
 			Text:       strings.TrimSpace(segment.Text),
 			Confidence: segment.GetConfidence(),
 		})
